Updates to authorization classes
