+++ src/main/java/auth/Operation.java
@@ -1,0 +1,11 @@
+package auth;
+
+/**
+ * @author Dave, Paul
+ */
+public enum Operation {
+    VIEW,
+    WRITE,
+    CREATE,
+    DELETE,
+}
