+++ src/main/java/auth/PrivilegeLevel.java
@@ -1,0 +1,11 @@
+package auth;
+
+/**
+ * @author Dave, Paul
+ */
+public enum PrivilegeLevel {
+    ADMIN,
+    COMMITTEE_LEADER,
+    COMMITTEE_MEMBER,
+    PARTICIPANT,
+}
