+++ src/main/java/auth/Who.java
@@ -1,0 +1,9 @@
+package auth;
+
+/**
+ * @author Dave, Paul
+ */
+public enum Who {
+    SELF,
+    OTHER,
+}
